:truck: rename: 비밀번호 변경 관련 컴포넌트 디렉토리 변경
